drivers: aw8697_haptic: fix -Wpointer-to-int-cast
../drivers/input/misc/aw8697_haptic/aw8697.c:4290:28: warning: cast to smaller integer type 'int' from 'u8 *' (aka 'unsigned char *') [-Wpointer-to-int-cast]
                        effect->pattern_length, (int)effect->pattern);
                                                ^~~~~~~~~~~~~~~~~~~~

+++ drivers/input/misc/aw8697_haptic/aw8697.c
@@ -4287,7 +4287,7 @@ static int aw8697_parse_dt_common(struct device *dev, struct aw8697 *aw8697,
 			       __func__);
 		}
 		printk("%s %d  effect->pattern_length=%d  effect->pattern=%d \n", __func__, __LINE__,
-			effect->pattern_length, (int)effect->pattern);
+			effect->pattern_length, (uintptr_t)effect->pattern);
 
 		effect->play_rate_us = config->play_rate_us;
 		rc = of_property_read_u32(child_node, "qcom,wf-play-rate-us",
